Updating result for baseline_api diff

+++ results/results_baseline_api_diff/gpt-4o-mini/0abf7148300f40a1da0538ab060552bca4a2f1d8/patches/patch1/updated/ReportBuilder.java
@@ -366,7 +366,7 @@ public class ReportBuilder
 			return;
 		}
 		
-		float lineWidth = border.getLineWidth();
+		float lineWidth = (float) border.getLineWidth();
 		textField.getLineBox().getPen().setLineWidth(lineWidth);
 		textField.getLineBox().getPen().setLineColor(border.getLineColor());
 		textField.getLineBox().getPen().setLineStyle(border.getLineStyle().getLineStyleEnum());
